refactor: rename of input component to index
