+++ src/components/common/MaxLengthNumberInput/MaxLengthNumberInput.tsx
@@ -1,5 +1,5 @@
 import { ChangeEvent, ElementRef, forwardRef, useState } from 'react'
-import { Input, InputProps } from '../Input/Input'
+import { Input, InputProps } from '../Input'
 import { css } from '@emotion/css'
 import { REGEXP } from '@/constants/regexp'
